# docs/swagger.yaml
basePath: /api/
definitions:
...
  models.CreateLocationRequest:
    properties:
      address:
        example: 123 Main St
        type: string
      capacity:
        example: 50000
        type: integer
      country:
        example: USA
        type: string
      stadium:
        example: National Stadium
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      message:
        example: An error occurred
        type: string
    type: object
...
  models.LocationResponse:
    properties:
      address:
        example: Wembley Park, London
        type: string
      capacity:
        example: 90000
        type: integer
      country:
        example: England
        type: string
      id:
        example: 101
        type: integer
      stadium:
        example: Wembley Stadium
        type: string
    type: object
  models.LocationsResponse:
    properties:
      locations:
        items:
          $ref: '#/definitions/models.LocationResponse'
        type: array
    type: object
...
  models.SuccessResponse:
    properties:
      message:
        example: Operation successful
        type: string
    type: object
  models.SuccessResponseCreate:
    properties:
      id:
        example: 1
        type: integer
      message:
        example: Object created successfully
        type: string
    type: object
...
  models.UpdateLocationRequest:
    properties:
      address:
        example: 456 Elm St
        type: string
      capacity:
        example: 75000
        type: integer
      country:
        example: Canada
        type: string
      stadium:
        example: Maple Leaf Stadium
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Simple API for managing ticket reservations.
  title: Ticket Reservation API
  version: "1.0"
paths:
...
  /locations:
    get:
      description: Retrieve a list of all locations.
      operationId: api.getLocations
      produces:
      - application/json
      responses:
        "200":
          description: List of locations
          schema:
            $ref: '#/definitions/models.LocationsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all locations
      tags:
      - locations
    put:
      consumes:
      - application/json
      description: Parse the payload and create a new location with provided dataset.
      operationId: api.createLocation
      parameters:
      - description: Payload to create a location
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateLocationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Location created successfully
          schema:
            $ref: '#/definitions/models.SuccessResponseCreate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create a new location.
      tags:
      - locations
  /locations/{id}:
    delete:
      description: Delete a location by its ID.
      operationId: api.deleteLocation
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event deleted successfully
          schema:
            $ref: '#/definitions/models.SuccessResponseCreate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete an existing location
      tags:
      - locations
    get:
      description: Retrieve a single location by ID.
      operationId: api.getLocation
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Location details
          schema:
            $ref: '#/definitions/models.LocationResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Retrieve a location.
      tags:
      - locations
    put:
      consumes:
      - application/json
      description: Update location details based on the provided payload.
      operationId: api.updateLocation
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      - description: Payload to update a location
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateLocationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Event updated successfully
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update an existing location
      tags:
      - locations
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"